rename: 컬럼 API 분리
